Split PATCHES
